Qol changes
General:
chore: added maybe_unused flags to prevent excessive warnings for prepared environment functions
chore: added explicit qualifiers partially
chore: added constexpr qualifiers partially for trivial math type operations
chore: some reformatting and grouping with visibility keywords
refactor: replaced c-style casts with static_casts partially

+++ src/Engine.Common/Hash/Murmur3.cpp
@@ -143,27 +143,27 @@ void hash::murmur3_x64_128(const void* data_, const u32 length_, const u64 seed_
 	u64 k2 = 0;
 
 	switch (length_ & 15) {
-		case 15: k2 ^= ((u64)tail[14]) << 48;
-		case 14: k2 ^= ((u64)tail[13]) << 40;
-		case 13: k2 ^= ((u64)tail[12]) << 32;
-		case 12: k2 ^= ((u64)tail[11]) << 24;
-		case 11: k2 ^= ((u64)tail[10]) << 16;
-		case 10: k2 ^= ((u64)tail[9]) << 8;
-		case 9: k2 ^= ((u64)tail[8]) << 0;
+		case 15: k2 ^= static_cast<u64>(tail[14]) << 48;
+		case 14: k2 ^= static_cast<u64>(tail[13]) << 40;
+		case 13: k2 ^= static_cast<u64>(tail[12]) << 32;
+		case 12: k2 ^= static_cast<u64>(tail[11]) << 24;
+		case 11: k2 ^= static_cast<u64>(tail[10]) << 16;
+		case 10: k2 ^= static_cast<u64>(tail[9]) << 8;
+		case 9: k2 ^= static_cast<u32>(tail[8]) << 0;
 
 			k2 *= c2;
 			k2 = ROTL64(k2, 33);
 			k2 *= c1;
 			h2 ^= k2;
 
-		case 8: k1 ^= ((u64)tail[7]) << 56;
-		case 7: k1 ^= ((u64)tail[6]) << 48;
-		case 6: k1 ^= ((u64)tail[5]) << 40;
-		case 5: k1 ^= ((u64)tail[4]) << 32;
-		case 4: k1 ^= ((u64)tail[3]) << 24;
-		case 3: k1 ^= ((u64)tail[2]) << 16;
-		case 2: k1 ^= ((u64)tail[1]) << 8;
-		case 1: k1 ^= ((u64)tail[0]) << 0;
+		case 8: k1 ^= static_cast<u64>(tail[7]) << 56;
+		case 7: k1 ^= static_cast<u64>(tail[6]) << 48;
+		case 6: k1 ^= static_cast<u64>(tail[5]) << 40;
+		case 5: k1 ^= static_cast<u64>(tail[4]) << 32;
+		case 4: k1 ^= static_cast<u64>(tail[3]) << 24;
+		case 3: k1 ^= static_cast<u64>(tail[2]) << 16;
+		case 2: k1 ^= static_cast<u64>(tail[1]) << 8;
+		case 1: k1 ^= static_cast<u64>(tail[0]) << 0;
 
 			k1 *= c1;
 			k1 = ROTL64(k1, 31);
